Added optional ussage of java int instead of BigInteger

+++ src/me/enerccio/sp/utils/Utils.java
@@ -22,7 +22,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.lang.reflect.Array;
-import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -30,13 +29,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Stack;
-
-import org.antlr.v4.runtime.ANTLRInputStream;
-import org.antlr.v4.runtime.BaseErrorListener;
-import org.antlr.v4.runtime.CommonTokenStream;
-import org.antlr.v4.runtime.RecognitionException;
-import org.antlr.v4.runtime.Recognizer;
-import org.antlr.v4.runtime.misc.ParseCancellationException;
 
 import me.enerccio.sp.compiler.PythonBytecode;
 import me.enerccio.sp.interpret.PythonExecutionException;
@@ -60,6 +52,13 @@ import me.enerccio.sp.types.sequences.ListObject;
 import me.enerccio.sp.types.sequences.SimpleIDAccessor;
 import me.enerccio.sp.types.sequences.StringObject;
 import me.enerccio.sp.types.sequences.TupleObject;
+
+import org.antlr.v4.runtime.ANTLRInputStream;
+import org.antlr.v4.runtime.BaseErrorListener;
+import org.antlr.v4.runtime.CommonTokenStream;
+import org.antlr.v4.runtime.RecognitionException;
+import org.antlr.v4.runtime.Recognizer;
+import org.antlr.v4.runtime.misc.ParseCancellationException;
 
 public class Utils {
 
@@ -137,7 +136,7 @@ public class Utils {
 
 		if (aType == Float.class || aType == float.class) {
 			if (datum instanceof RealObject)
-				return ((RealObject)datum).floatValue();
+				return ((RealObject)datum).doubleValue();
 			else
 				return asJavaObject(aType, PythonInterpret.interpret.get().executeCall("float", datum));
 		}
